Delete and retry stuck locks

+++ honeyflare/locks.py
@@ -1,7 +1,8 @@
 from io import BytesIO
+from datetime import datetime, timezone
 
 from google.cloud import storage
-from google.api_core.exceptions import PreconditionFailed
+from google.api_core.exceptions import PreconditionFailed, NotFound
 
 from .exceptions import FileLockedException
 
@@ -13,6 +14,8 @@ storage.blob._MULTIPART_URL_TEMPLATE += '&ifGenerationMatch=0'
 storage.blob._RESUMABLE_URL_TEMPLATE += '&ifGenerationMatch=0'
 
 EMPTY_FILEOBJ = BytesIO()
+# This is the maximum support timeout by Google Cloud Functions
+MAX_EXECUTION_TIME_SECONDS = 540
 
 
 class GCSLock():
@@ -35,11 +38,37 @@ def lock(bucket, lock_name):
     blob = bucket.blob(lock_name)
     try:
         blob.upload_from_file(EMPTY_FILEOBJ)
+        return True
     except PreconditionFailed:
-        return False
-    return True
+        # Check that the lock was created recently (ie the creator function
+        # might still be running)
+        blob.reload()
+        lock_age = datetime.now(timezone.utc) - blob.time_created
+        if lock_age.total_seconds() > MAX_EXECUTION_TIME_SECONDS:
+            try:
+                blob.delete()
+            except NotFound:
+                # A parallel function might have just deleted the lock
+                pass
+
+            # Retry acuiring the lock
+            try:
+                blob.upload_from_file(EMPTY_FILEOBJ)
+                return True
+            except PreconditionFailed:
+                pass
+
+    return False
 
 
 def unlock(bucket, lock_name):
     blob = bucket.blob(lock_name)
-    blob.delete()
+    try:
+        blob.delete()
+    except NotFound:
+        # This can happen if another function thinks we have timed out but we
+        # are for some reason still running (which can happen if the instance
+        # that timed out is scheduled for more work after timing out for the
+        # first request) At this point both functions will
+        # have submitted the data, not much we can do to recover.
+        pass
